Serve status socket on a named pipe on Windows
Status commands were not available on Windows because the status
component was listening only on a UNIX socket, which is not available on
Windows. This change makes the status component listen on a named pipe
on Windows using Microsoft's go-winio package, which was already a
k0s dependency.

[PR 2897]

+++ pkg/component/status/status.go
@@ -23,6 +23,7 @@ import (
 	"net"
 	"net/http"
 	"os"
+	"runtime"
 	"strconv"
 	"time"
 
@@ -56,6 +57,10 @@ type certManager interface {
 
 var _ manager.Component = (*Status)(nil)
 
+var listen func(string) (net.Listener, error) // set by listen.go / listen_windows.go
+
+var listen listenFunc = net.Listen
+
 const defaultMaxEvents = 5
 
 // Init initializes component
@@ -83,7 +88,7 @@ func (s *Status) Init(_ context.Context) error {
 	}
 
 	removeLeftovers(s.Socket)
-	s.listener, err = net.Listen("unix", s.Socket)
+	s.listener, err = listen(s.Socket)
 	if err != nil {
 		s.L.Errorf("failed to create listener %s", err)
 		return err
@@ -95,6 +100,10 @@ func (s *Status) Init(_ context.Context) error {
 
 // removeLeftovers tries to remove leftover sockets that nothing is listening on
 func removeLeftovers(socket string) {
+	if runtime.GOOS == "windows" {
+		return
+	}
+
 	_, err := net.Dial("unix", socket)
 	if err != nil {
 		_ = os.Remove(socket)
